// packages/pieces/community/quickbooks-sandbox/src/lib/types.ts
export interface QuickbooksCurrencyRef {
    value: string;
    name?: string;
}

export interface QuickbooksMetaData {
    CreateTime?: string;
    LastUpdatedTime?: string;
}
// ...
export interface QuickbooksAccount {
    Id: string;
    Name: string;
    AccountType: string; // e.g., 'Bank', 'Credit Card', 'Accounts Receivable', 'Expense', 'Income'
    AccountSubType?: string;
    Classification?: string; // e.g., 'Asset', 'Liability', 'Equity', 'Revenue', 'Expense'
    Active?: boolean;
    CurrencyRef?: QuickbooksCurrencyRef;
    CurrentBalance?: number;
    domain?: string;
    sparse?: boolean;
    MetaData?: QuickbooksMetaData;
}